refactor: remove unnecessary components
