+++ nextjs/src/app/connections/page.tsx
@@ -1,4 +1,4 @@
-import Connections from '@/features/connections'
+import Connections from '@/features/connections/components/Connections'
 import React from 'react'
 
 const page = (): React.JSX.Element => <Connections />
